# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Budget Monitor API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cycles:
    get:
      operationId: listCycles
      tags: [cycles]
      summary: List all budget cycles
      responses:
        "200":
          description: List of cycles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetCycle"

    post:
      operationId: createCycle
      tags: [cycles]
      summary: Create a budget cycle (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCycleBody"
      responses:
        "201":
          description: Cycle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCycle"
...
  /cycles/{cycleId}:
    put:
      operationId: updateCycle
      tags: [cycles]
      summary: Update a budget cycle (admin only)
      parameters:
        - name: cycleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCycleBody"
      responses:
        "200":
          description: Cycle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCycle"
  /cycles/{cycleId}/activate:
    post:
      operationId: activateCycle
      tags: [cycles]
      summary: Activate a budget cycle (deactivates all others)
      parameters:
        - name: cycleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cycle activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCycle"

  # ALLOCATIONS
  /allocations:
    get:
      operationId: listAllocations
      tags: [allocations]
      summary: List allocations (filtered by user's scope)
      parameters:
        - name: cycleId
          in: query
          required: false
          schema:
            type: integer
        - name: sectorId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, active, revoked, exhausted]
      responses:
        "200":
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AllocationWithDetails"

    post:
      operationId: createAllocation
      tags: [allocations]
      summary: Allocate funds to a sector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationBody"
      responses:
        "201":
          description: Allocation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationWithDetails"
        "400":
          description: Validation error (e.g. insufficient balance)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /allocations/{allocationId}:
    get:
      operationId: getAllocation
      tags: [allocations]
      summary: Get a specific allocation
      parameters:
        - name: allocationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Allocation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationWithDetails"

  # REVOCATIONS
  /allocations/{allocationId}/revoke:
    post:
      operationId: revokeAllocation
      tags: [revocations]
      summary: Revoke an allocation
      parameters:
        - name: allocationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeAllocationBody"
      responses:
        "200":
          description: Allocation revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationWithDetails"
        "400":
          description: Already revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # DASHBOARD SUMMARIES
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products in the catalog
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # PURCHASE ORDERS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, ceo, ministry_head, department_head, viewer]
        sectorId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
...
    Sector:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        parentId:
          type: integer
          nullable: true
        depth:
          type: integer
        responsibleUserId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - code
        - depth
        - isActive
        - sortOrder
        - createdAt
...
    BudgetCycle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        totalBudget:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
        totalAllocated:
          type: number
        totalRevoked:
          type: number
        availableBalance:
          type: number
        utilizationPct:
          type: number
      required:
        - id
        - name
        - totalBudget
        - startDate
        - endDate
        - isActive
        - createdBy
        - createdAt
    CreateCycleBody:
      type: object
      properties:
        name:
          type: string
        totalBudget:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
      required:
        - name
        - totalBudget
        - startDate
        - endDate
    UpdateCycleBody:
      type: object
      properties:
        name:
          type: string
        totalBudget:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
    Allocation:
      type: object
      properties:
        id:
          type: integer
        budgetCycleId:
          type: integer
        fromSectorId:
          type: integer
          nullable: true
        toSectorId:
          type: integer
        allocatedBy:
          type: integer
        amount:
          type: number
        comment:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, active, revoked, exhausted]
        allocatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - budgetCycleId
        - toSectorId
        - allocatedBy
        - amount
        - status
        - allocatedAt
        - createdAt
    AllocationWithDetails:
      allOf:
        - $ref: "#/components/schemas/Allocation"
        - type: object
          properties:
            fromSector:
              $ref: "#/components/schemas/Sector"
              nullable: true
            toSector:
              $ref: "#/components/schemas/Sector"
            allocatedByUser:
              $ref: "#/components/schemas/User"
            revocation:
              $ref: "#/components/schemas/Revocation"
              nullable: true
    CreateAllocationBody:
      type: object
      properties:
        budgetCycleId:
          type: integer
        fromSectorId:
          type: integer
          nullable: true
        toSectorId:
          type: integer
        amount:
          type: number
        comment:
          type: string
      required:
        - budgetCycleId
        - toSectorId
        - amount
    RevokeAllocationBody:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
    Revocation:
      type: object
      properties:
        id:
          type: integer
        allocationId:
          type: integer
        revokedBy:
          type: integer
        reason:
          type: string
        revokedAt:
          type: string
          format: date-time
        revokedByUser:
          $ref: "#/components/schemas/User"
      required:
        - id
        - allocationId
        - revokedBy
        - reason
        - revokedAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        unitPrice:
          type: number
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - unit
        - unitPrice
        - isActive
        - sortOrder
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        unitPrice:
          type: number
        description:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - name
        - category
        - unit
        - unitPrice
...